requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.1.0
Brotli==1.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
discord.py==2.3.2
ffmpeg==1.4
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
mutagen==1.47.0
pycparser==2.21
pycryptodomex==3.19.0
PyNaCl==1.5.0
websockets==12.0
yarl==1.9.2
yt-dlp==2023.10.13
